backend/pyproject.toml
[project]
name = "clipabit-backend"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "python-multipart",
    "python-dotenv",
    "ffmpeg-python",
    "opencv-python-headless",
    "numpy",
    "modal",
    "pinecone",
    "torch",
    "pillow",
    "transformers",
    "scenedetect",
    "boto3",
    "torchvision",
    "firebase-admin>=7.1.0",
]

[project.scripts]
dev = "cli:serve_all"
server = "cli:serve_server"
search = "cli:serve_search"
processing = "cli:serve_processing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli.py"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio>=1.3.0",
    "pytest-cov",
    "pytest-mock",
]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
pythonpath = ["."]
python_files = ["test_*.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["preprocessing", "database", "models", "embeddings"]
omit = ["tests/*", "utils/*", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
precision = 2
